Add warning message on failed search Reason: Failed search -> A search which yielded no results. Before this commit, a failed search would redirect the user back to the /browse?page=1 route without any indication that the search has failed.
How this change addresses the issue:
A displayError boolean in the Browse component's state will toggle
depending on the data returned in the callback of the first fetch
request. If true, the message is appended as the first child of the
Segment element.

Side Effects:
The message may appear if the user types too slow, which causes the
fetch request to occur prematurely, which may confuse the user.